requirements.txt
paramiko==2.4.1
pexpect==4.5.0
PyYAML==3.12
progressbar==2.3
tqdm==4.19.6
aexpect==1.4.0